refactor(releases): 抽出 useWatchedSourcesSync,修复设置弹窗过度重渲染
ReleaseSourceSettingsModal 只需要 syncWatchedSources / isSyncingWatchedSources
两项,却经 useReleaseTimelineActions 订阅了 selectReleaseTimelineState 全部
35 个字段,任一 Release 状态变更都会让它重渲染。把同步逻辑整体搬入窄订阅
hook useWatchedSourcesSync(仅订阅 githubToken/language/setReleaseSource-
Repositories 三个单值 selector),modal 改为直接消费;同步用例测试随逻辑
迁至 useWatchedSourcesSync.test.tsx。

ReleaseTimeline 不消费 sync 相关字段,行为不受影响;同步回调逐字保留。
