+++ src/features/releases/hooks/useReleaseTimelineActions.ts
@@ -12,7 +12,6 @@ import {
   getSourcesForReleaseRepository,
   normalizeRepoKey,
   releaseBelongsToResolvedSources,
-  repositoryToCustomReleaseRepository,
   resolveReleaseSources,
   STARRED_RELEASE_SOURCE_ID,
   WATCH_CUSTOM_RELEASE_SOURCE_ID,
@@ -25,13 +24,9 @@ import { findReleasesWithChangedAssets } from '../../../utils/releaseAssets';
  */
 export const useReleaseTimelineActions = () => {
   const state = useAppStore(useShallow(selectReleaseTimelineState));
-  // 共享 selector 未收录的 store 写入用独立单值 selector 获取（B1）；
-  // store action 引用稳定，单值订阅不会引入多余重渲染。
-  const setReleaseSourceRepositories = useAppStore((s) => s.setReleaseSourceRepositories);
   const { toast, confirm } = useDialog();
   const [lastRefreshTime, setLastRefreshTime] = useState<string | null>(null);
   const [isMarkingAllRead, setIsMarkingAllRead] = useState(false);
-  const [isSyncingWatchedSources, setIsSyncingWatchedSources] = useState(false);
   const t = useCallback((zh: string, en: string) => state.language === 'zh' ? zh : en, [state.language]);
 
   const handleRefresh = useCallback(async () => {
@@ -194,51 +189,13 @@ export const useReleaseTimelineActions = () => {
     toast(t('已取消订阅该仓库的 Release。', 'Unsubscribed from repository releases.'), 'success');
   }, [state, t, toast, confirm]);
 
-  // 同步 Watch 仓库为 Release 来源（原 ReleaseSourceSettingsModal →
-  // WatchCustomReleaseSyncPanel.handleSync）：无 confirm、无 forceSync、无 Abort。
-  const syncWatchedSources = useCallback(async () => {
-    if (!state.githubToken || isSyncingWatchedSources) return;
-
-    setIsSyncingWatchedSources(true);
-    try {
-      const githubApi = new GitHubApiService(state.githubToken);
-      // 只拉 /user/subscriptions（含私有仓）。/users/{login}/subscriptions 已被 GitHub 改为
-      // 恒定返回 204 空响应体，且其结果本就是前者的公开子集，并行合并只会拖垮整个同步。
-      const watchedRepos = await githubApi.getAllWatchedRepositories();
-      // hiddenByRepo 在 await 之后从最新 state 读取：同步期间用户仍可在设置面板
-      // 切换 release_hidden，旧快照会在 setReleaseSourceRepositories 时覆盖该修改。
-      const hiddenByRepo = new Map(
-        useAppStore.getState().releaseSourceSettings.watchCustomReleaseRepos.map(repo => [normalizeRepoKey(repo.full_name), repo.release_hidden]),
-      );
-      const sourceRepos = watchedRepos.map(repo => ({
-        ...repositoryToCustomReleaseRepository(repo, WATCH_CUSTOM_RELEASE_SOURCE_ID),
-        release_hidden: hiddenByRepo.get(normalizeRepoKey(repo.full_name)) || undefined,
-      }));
-      setReleaseSourceRepositories(WATCH_CUSTOM_RELEASE_SOURCE_ID, sourceRepos);
-      toast(
-        t(
-          `已同步 ${sourceRepos.length} 个 Watch 仓库。`,
-          `Synced ${sourceRepos.length} Watch repositories.`
-        ),
-        'success'
-      );
-    } catch (error) {
-      console.error('Failed to sync watched repositories:', error);
-      toast(t('同步 Watch 仓库失败，请检查网络或 Token 权限。', 'Failed to sync Watch repositories. Check network or token permissions.'), 'error');
-    } finally {
-      setIsSyncingWatchedSources(false);
-    }
-  }, [state, isSyncingWatchedSources, setReleaseSourceRepositories, t, toast]);
-
   return {
     ...state,
     lastRefreshTime,
     isMarkingAllRead,
-    isSyncingWatchedSources,
     handleRefresh,
     handleMarkAllRead,
     handleUnsubscribeRelease,
-    syncWatchedSources,
   };
 };
